requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.0=py39h07f9747_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=py_1
brotlipy=0.7.0=py39h38d8fee_1001
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py39hf3d152e_1
cffi=1.15.0=py39hd667e15_1
chardet=3.0.4=py39h079e4ff_1008
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.0=py39hf3d152e_0
clip=1.0=pypi_0
cryptography=35.0.0=py39hbca0aa6_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cython=0.29.28=pypi_0
debugpy=1.2.0=py39h41458e0_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.31.2=pypi_0
freetype=2.11.0=h70c0345_0
ftfy=6.0.3=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.44.0=pypi_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py39hf3d152e_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.8.0=py39hef51801_0
ipython=7.31.1=py39hf3d152e_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=py39hf3d152e_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=4.4.0=pypi_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.1=py39hf3d152e_1
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.4.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=4.7.5=py39h07f9747_2
nbclient=0.5.10=pypi_0
nbconvert=6.4.1=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.8=pypi_0
numpy=1.22.3=pypi_0
oauthlib=3.2.0=pyhd8ed1ab_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
ortools=9.3.10497=pypi_0
packaging=21.3=pypi_0
pandas=1.4.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.26=pyha770c72_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py39hf3d152e_4
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.2=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=20.0.0=py39h25affbc_1
readline=8.1.2=h7f8727e_1
regex=2022.1.18=pypi_0
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
terminado=0.13.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.10.2=py39_cu102
torchvision=0.11.3=py39_cu102
tornado=6.1=py39hbd71b63_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.9=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
werkzeug=2.1.0=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py39hbd71b63_0
zeromq=4.3.3=h58526e2_3
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0